delete history v3.1
